Added metrics example to examples

+++ src/main/stats.cc
@@ -109,11 +109,11 @@ static Local<Object> build_node_stats(as_node_stats *node)
 	Nan::Set(stats, Nan::New("pipelineConnections").ToLocalChecked(),
 			 build_conn_stats(&node->pipeline));
 	Nan::Set(stats, Nan::New("errorCount").ToLocalChecked(),
-			 Nan::New(node->error_count));
+			 Nan::New(static_cast<double>(node->error_count)));
 	Nan::Set(stats, Nan::New("timeoutCount").ToLocalChecked(),
-			 Nan::New(node->timeout_count));
+			 Nan::New(static_cast<double>(node->timeout_count)));
 	Nan::Set(stats, Nan::New("keyBusyCount").ToLocalChecked(),
-			 Nan::New(node->key_busy_count));
+			 Nan::New(static_cast<double>(node->key_busy_count)));
 	return stats;
 }
